Added strict type checking (#14)
* Added strict type checking

+++ tests/test_micro_cluster.py
@@ -52,9 +52,9 @@ class TestMicroCluster(unittest.TestCase):
         lambd = 0.5
 
         # Feature-arrays
-        x1 = np.array([1, 2]).reshape((1, 2))
-        x2 = np.array([4, 5]).reshape((1, 2))
-        x3 = np.array([0, 0]).reshape((1, 2))
+        x1 = np.array([1, 2], dtype=np.float32).reshape((1, 2))
+        x2 = np.array([4, 5], dtype=np.float32).reshape((1, 2))
+        x3 = np.array([0, 0], dtype=np.float32).reshape((1, 2))
 
         # Time values
         t1 = 1.0
